add distance signs

+++ src/model.rs
@@ -15,8 +15,8 @@ pub enum GameState {
 #[derive(Resource)]
 pub struct GameData {
     pub game_state: GameState,
-    pub previous_score: usize,
-    pub current_score: usize,
+    pub previous_score: f32,
+    pub current_score: f32,
     pub velocity: f32,
 }
 
@@ -42,8 +42,8 @@ impl Plugin for Model {
     fn build(&self, app: &mut App) {
         app.insert_resource(GameData {
             game_state: GameState::WaitingForStart,
-            previous_score: 0,
-            current_score: 0,
+            previous_score: 0.,
+            current_score: 0.,
             velocity: INITIAL_VELOCITY
         }).add_observer(handle_model_reset);
     }
@@ -55,7 +55,7 @@ fn handle_model_reset(
 )
 {
     game.previous_score = game.current_score;
-    game.current_score = 0;
+    game.current_score = 0.;
     game.velocity = INITIAL_VELOCITY;
     game.game_state = GameState::WaitingForStart;
 }
